pyproject.toml
[project]
name = "perp-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ccxt",
    "python-dotenv",
    "rich",
]

[dependency-groups]
dev = [
    "pyinstrument",
]
